Use RetryFor instead of sleep (#7384)
Some of our integration tests rely on client (or server) workloads to
test Linkerd behaves accordingly. Because of this, some tests are prone
to transient errors, commonly associated with pods being in an
initialization phase; either the pod itself is not marked ready, or
traffic hasn't started flowing.

A common pattern we've used throughout some of these tests is to block
the execution of the test with time.Sleep. This gives the workloads
enough time to start sending or receiving traffic. Instead of blocking
execution, it would be better to make use of RetryFor with a timeout.
This way, pods may start up before the timer completes.

Signed-off-by: Matei David <matei@buoyant.io>