Only set privileged: true if really needed (#7401)
The recently introduced runAsRoot override in #7348 for supporting Docker runtime is a bit to permissive and will set privileged: true although that is only needed if .proxyInit.closeWaitTimeoutSecs is set. This prevents proxy container from running in namespaces which are not entitled to run privileged containers (e.g. due to PodSecurityPolicies).

Through this change, we will only set privileged: true if it is really needed (i.e. .proxyInit.closeWaitTimeoutSecs is set).

Signed-off-by: Alexander Berger <alex-berger@gmx.ch>