Set up helm charts versioning
Implements part of #7405

Separates `version` from `appVersion` in both `linkerd-crds` and
`linkerd-control-plane` charts.

In `linkerd-crds` embeds `version` as a label in the CRDs.

In `linkerd-control-plane`, embeds `linkerd-crds`'s `version` as a new
entry in the `linkerd-config` ConfigMap, to use in a new `linkerd check`
that will be implemented in a separate PR.