Relax extension checks (#7406)
Fixes #7385

If an extension pod is not running, this does not necessarily mean that the extension is unable to function since there may be other replicas which are healthy.  We want to be conservative about only blocking CLI functionality when we are absolutely sure that it won't work.  Thus, we relax these checks to the warning level so that they don't block access to CLI functionality.

Signed-off-by: Alex Leong <alex@buoyant.io>