Use CheckPodsRunning in Linkerd healthcheck (#7452)
`validateDataPlanePods` and `CheckPodsRunning` both check that for a given list of Pods, each Pod is running and has a `linkerd-proxy` container that is ready.

Aside from the error message returned when there are no Pods given, these functions do the same thing. This change merges the two in favor for `CheckPodsRunning` which is used by other packages.

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>