Proper mirroring of empty Endpoints in multicluster (#7520)
* Proper mirroring of empty Endpoints in multicluster

(This is required to support the linkerd-failover effort).

Pod readiness is reflected in the emptiness of the associated Endpoints (check
the comment on the function `isEmptyService` for what emptiness means exactly).
Currently, target empty Endpoints mirrors always have as `addresses` the target
Gateway address, regardless of the emptiness of the target Service. This PR
attempts to properly mirror that emptiness.

We expand the Endpoints watcher set in `cluster_watcher.go` to not only account
for EP from headless services, but all exported services:
- `Add` events are only considered for exported headless services.
- `Update` events are considered for all exported services:
  - If the service is headless, the old handler is used
    (`createOrUpdateHeadlessEndpoints`) (*)
  - If the service is not headless, a new handler is used, that makes sure the
    emptiness of the target EP is reflected on the source EP.

During `repairEndopints` there's also additional logic to check the emptiness
of the target EP and reflect that in the source EP.

Bonus: I bumped the service-mirror log level to debug in integration
tests.

I also noticed we could make `cluster_watcher.go` simpler by moving the
headless logic into a separate file, but that can be accomplished afterwards in
a separate PR.

(cherry picked from commit c2eac91f508dc632de5135e5deb98e39136067f6)
Signed-off-by: Oliver Gould <ver@buoyant.io>