Properly version Helm charts for extensions (#7568)
This follows up the the versioning scheme used for `linkerd-crds` and
`linkerd-control-plane` but for extension charts. So no longer will
a chart's `version` just mimic `appVersion`'s value, but instead it'll
change depending on how the chart changes on each realease.

Unfortunately this means a bit more of manual work during releases, but
we should be able to automate some of it at some point.