Reset extensions Helm chart versions at 30.0.0-edge
#7568 divorced Helm chart versions (`version`) from `appVersion` in the
extension charts, following suit to what we did for the core
`linkerd-crds` and `linkerd-control-plane` charts. But since the charts
are still the same, their `version` has to be greater than previous
releases, in order to be able to upgrade to releases using the new
version scheme. Therefore, we're resetting version at `30.0.0-edge`.