Add a devcontainer configuration (#7583)
VS Code [devcontainers][dc] provide a way to define a reproducible,
customizable developer environment. This should lower the barrier to
entry for contributors and reduce the maintenance burden for existing
developers, as a docker daemon is now basically all that's needed to get
up and running.

This change adds a devcontainer for the Linkerd2 project. Its container
starts in the host's network so that it can easily access k3d clusters
running on the host's docker daemon.

[dc]: https://code.visualstudio.com/docs/remote/containers

Signed-off-by: Oliver Gould <ver@buoyant.io>
(cherry picked from commit f33bb7beed734e433296bbb1b68d065da11127bf)
Signed-off-by: Oliver Gould <ver@buoyant.io>