Remove `dashmap` from policy-controller indexer
`DashMap` is a concurrent hashmap implementation that uses unsafe code.
The policy controller uses it to facilitate policy lookups in the gRPC
API. In practice, this map need not be particularly low-latency or
high-throughput: updates and lookups are relatively low-volume. On the
other hand, the dashmap necessarily uses unsafe code, and poses a
security/safety risk (e.g., RUSTSEC-2022-02). It seems wiser to use a
simpler implementation until we have evidence that we need a more
complex solution.

This change replaces uses of `DashMap` with a simple
`RwLock<HashMap<...>>`. No new dependencies are introduced. The
`dashmap` dependency will not be dropped completely until
kube-rs/kube-rs#781 is addressed.

Addresses #7606.

Signed-off-by: Oliver Gould <ver@buoyant.io>