Parametrized datasource in grafana dashboards, better script handling (#7603)
Added the `DS_PROMETHEUS` parameter in all the Grafana dashboard
definitions. When importing a definition into a Grafana Cloud instance
for example, the import form will allow selecting the datasource from
  the currently available. OTOH when using an in-cluster instance such
  as one installed through the Grafana helm chart, the parameter gets
  overridden with the `values.yaml` entry
  `dashboards.default.{name}.datasource`.

Also, the javascript snippet used in the dashboard definitions for
checking for the latest linkerd version has been wrapped around a hidden
div. This avoids showing the script itself when it gets escaped when
importing the definition into Grafana Cloud.