Service mirror controllers should not create namespaces (#7597)
Fixes #7565

When the service mirror controller detects an exported service in the target cluster, it attempts to create a mirror service for it in the source cluster in a namespace with the same name.  If that namespace does not exist in the source cluster, the service mirror controller will create it first.  This is somewhat unexpected behavior since the service mirror controller should not be in charge of administering cluster level resources such as namespaces.

We update the service mirror controller to instead log a warning and skip mirroring in this case.  Users who want to export a service should first ensure that the namespace exists in the source cluster.

Signed-off-by: Alex Leong <alex@buoyant.io>