Change erroneous 'minutes' to 'seconds' in window
The metrics api helper in our web component contains a set of valid time
windows for the collection and display of metrics. Although this is
potentially not actively used in the dashboard itself, the library
contains an erroneous measure whereby 10s are expressed as minutes.

This is a quick fix to change the seconds field to the appropriate
scale.

Signed-off-by: Matei David <matei@buoyant.io>