Remove jaeger existence check (#7750)
Fixes #7663

The `linkerd jaeger check` command checks that jaeger and the opencensus-collector are installed and that all of the associated resources such as configmaps and service accounts exist.  However, the linkerd-jaeger extension allows users to skip installing jaeger and the oc-collector by using the `jaeger.enabled=false` and `collector.enabled=false` values respectively.  This is to support the BYO jaeger or BYO collector use-cases.  However, if either jaeger or the collector is not installed, check will fail.

Since it is valid for these resources to not exist, we remove these checks.  There isn't an easy way to know at runtime which values the extension was installed with, so it is difficult to know if these resources should exist or not.  

Signed-off-by: Alex Leong <alex@buoyant.io>