Allow integration tests to run in parallel
Go's test runner (`go test`) can be non-deterministic with the order in
which it runs the tests. This is due to the way package paths are
handled in Go, and due to its parallelism; tests in Go seem to be always
run in parallel, but the specifics here differ depending on the
available CPU.

Normally, we'd expect our tests to be run in a
predictable order, starting with `1_install_test`. To force this
behaviour, we rely on a `-p 1` argument that forces the runner to run
all tests sequentially. If we remove this, tests are prone to being run
in parallel and they will start executing before the control plane is in
a ready state.

We can take advantage of parallelism here to get better timing on our
tests, however, we need to block the start of each test until the
control plane (or extension) pods are ready. In each `TestMain`, we
block until the pods are ready.

Signed-off-by: Matei David <matei@buoyant.io>