devcontainer: Pull rust installation from docker
We currently install rustup manually and then install a rust toolchain;
but this is slow and serial.

This change moves rust environment provisioning into a separate build
stage that fetches a rust environment in parallel with other build
steps. The rust environment is copied over towards the end of the build.

Signed-off-by: Oliver Gould <ver@buoyant.io>