Re-enable tracing test (#7840)
The tracing integration test seems to reliably pass now.  It's unclear exactly what the cause of the flakiness was, but it is possible that recent refactors in the integration tests have alleviated it.  We re-enable this test for now.  We also add some additional checks to verify the annotations on the running pods to ensure they have been processed by the jaeger injector.  This may help narrow down the source of flaky failures in the future.

Signed-off-by: Alex Leong <alex@buoyant.io>