Remove usage of controllerImageVersion values field
We have removed the option of configuring the controller image version
through a dedicated CLI flag (no more `--controller-image-version`). All
of our controller images will now mirror `linkerdVersionValue`, and that
should be the only way of overwriting the images in manifests. The
exception here is the policy controller which seems to have its own
field (no corresponding flag, it still has to be set through `--set`).

This change removes the unused `controllerImageVersion` field, first
from the tests, and then from the actual chart values structure. Note
that at this point in time, it is impossible to use
`--controller-image-version` through Helm, yet it still seems to be
working for the CLI.

As part of this change:

* We configure the charts to use `linkerdVersionValue` instead of
  `controlPlaneImageVersion` (or default to it where appropriate).
* We add the stringslicevar flag (i.e `--set`) to the flagset we use in
  upgrade tests. This means instead of testing value overrides through a
  dedicated flag, we can now make use of `--set` in upgrade tests. We
  first set the linkerdVersionValue in the install option and then
  override the policy controller image version and the linkerd
  controller image version to test flags work as expected.
* We remove hardcoded values from healthcheck test.
* We remove field from chart values struct.

Signed-off-by: Matei David <matei@buoyant.io>