Fix error handling in wehbook server [TOB-LNKD-9] (#7882)
When the webhook server decodes a request's JSON payload, it may try to
use a nil value when handling the error. Furthermore, if the JSON
payload has a `nil` `Request` value, it may attempt to dereference the
value.

This change improves the webhook server's error handling to return a
`400 Bad Request` status if either of these cases are encountered.

Signed-off-by: Oliver Gould <ver@buoyant.io>