Test for validating proxy can't be used as an open relay (#7884)
* Test for validating proxy can't be used as open relay

Closes #7697

We setup two injected bb servers and an uninjected curl client. One of
the servers is injected in ingress mode. The curl client tries to
connect to it to the port 4140 while setting the `l5d-dst-override` header
to attempt a relay to the other server.

The first test shows this doesn't result in a relay. The second test
does the same, but changing LINKERD2_PROXY_OUTBOUND_LISTEN_ADDR in the
injected server proxy, making it listen on all addresses. In that case,
the relay succeeds.