upgrade web/ to use protojson (#7995)
As part of #7881 it is recommended we upgrade to use `protojson` instead of `jsonPb`. This required a few additional changes in the code since the signature of `Marshal` changed so I've split this out into a separate change.

`Marshal` now returns the resulting bytes instead of taking an interface to write those bytes into. This means we have to handle the error separately and then write the JSON into the given writer.

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>