empty mirror service endpoints when gateway is down (#8022)
Closes linkerd/linkerd-failover#63

When the gateway in a target cluster is down, the endpoints for mirror services on a source cluster do not reflect that the target cluster will be unreachable. To fix this, the endpoints for mirror services should transition their addresses to not ready.

Currently, every connected cluster has a probe worker and a cluster watcher. The probe worker is responsible for probing the gateway on the target cluster, and the cluster watcher is responsible for watching for events on the target cluster and making necessary changes on the source cluster. The probe worker and cluster watcher do not communicate in any way.

This change introduces the ability for the probe worker to update the cluster watcher when the gateway has a change in liveness. When a gateway's liveness status changes, the cluster watcher is updated and triggers an endpoint repair. If the gateway is now down, the endpoint repair transitions all endpoints for mirror services to not ready. If the gateway is alive, the endpoint repair will ensure that all endpoints for mirror services point to the ready gateway address.

There is a possibility that a gateway is unreachable but the local cluster can still connect to the remote cluster. When this is the case, we still honor the probe's liveness status. This means for example that if a remote service is created when the gateway is down, the local cluster may still see that update (via the cluster watcher), but the endpoints that it creates for the mirror service are not ready.

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>