split creating and updating mirror endpoints (#8075)
#8022 introduced `createOrUpdateEndpoints` which creates or updates mirror endpoints in multicluster installations. The reason that creating and updating endpoints was put behind a single function is so that we could consolidate the `gatewayAlive` check down to a single location. This is good because we know that we aren't missing any places in the service mirror code where we should be checking `gatewayAlive`.

This isn't ideal though for updating endpoints because it always involves an API `Endpoints(...).Get(...)` to satisfy the `Create` case. If `Get` fails, we know we have to create; if it does not fail we _ignore_ the returned endpoints and instead `Update` with the given endpoints. While this isn't a huge bottleneck, it does scale with the number of services that are being mirrored and is easy to skip.

We always know whether we want to update or create; we never are in a case of falling back to create endpoints. Therefore, we can split this function into separate `Create` and `Update` API calls and just make sure to have the same `gatewayAlive` check at the start of each.

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>
(cherry picked from commit c78b4259053bd03fec01ee0ff3a16a577d958c9c)
Signed-off-by: Oliver Gould <ver@buoyant.io>